RAG/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.0",
    "bs4>=0.0.2",
    "faiss-cpu>=1.13.2",
    "groq>=0.37.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "numpy>=2.4.2",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "wikipedia>=1.4.0",
]
